% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_experiment_tables.R
\name{add_experiment_tables}
\alias{add_experiment_tables}
\title{Add \code{composition} and \code{experimetal_design} table to db.}
\usage{
add_experiment_tables(
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  composition = NULL,
  experimetal_design = NULL,
  light_decline_schedule = NULL,
  immigration_schedule = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{composition}{fully qualified path to the \code{csv} file with the composition data}

\item{experimetal_design}{fully qualified path to the \code{csv} file with the experimental design table}

\item{light_decline_schedule}{fully qualified path to the \code{csv} file with the light decline dates}

\item{immigration_schedule}{fully qualified path to the \code{csv} file with the immigration event dates}

\item{overwrite}{if \code{TRUE}, the existing tables will be overwritten. There is actually no need for this.}
}
\description{
Add \code{composition} and \code{experimetal_design} table to db.
}
