% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_extracted_csv_to_db.R
\name{add_extracted_csv_to_db}
\alias{add_extracted_csv_to_db}
\title{Title}
\usage{
add_extracted_csv_to_db(
  db,
  csv_file = "Morph_mvt_non_cropped.csv",
  measure = "bemovi.mag.25",
  tn_postfix = "non_cropped",
  archive_dir = "/Users/rainerkrug/Duck/LEEFSwift3",
  copy_locally_first = FALSE
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database.Must be set!}

\item{csv_file}{name of the csv files to be imported}

\item{measure}{measure}

\item{tn_postfix}{postfix for the table name,}

\item{archive_dir}{directory containes the archive. It contains the following directories:
- LEEF.archived.data
- LEEF.archived.data_segments
- LEEF.backend.data
- LEEF.backend.data_segments}

\item{copy_locally_first}{}
}
\description{
Title
}
