% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_db.R
\name{add_to_db}
\alias{add_to_db}
\title{Add data to RRD database into existing table}
\usage{
add_to_db(fn, dbname, table)
}
\arguments{
\item{fn}{\code{rds} file containing data to be added. The name has to be as expected for the additor function.}

\item{table}{name of the table the data should be added to}

\item{db}{RRD sqlite databaes}
}
\value{
\code{TRUE} if the data has been added, \code{FALSE} otherwise
}
\description{
Add data to RRD database into existing table
}
