% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen_deposit.R
\name{all_archives}
\alias{all_archives}
\title{Get all measurement_method / timestamp combinations in the archive}
\usage{
all_archives(archive_dir = "~/Duck/LEEFSwift3", stage)
}
\arguments{
\item{archive_dir}{base directory of the data archive}

\item{stage}{either \code{"pre_processed"} or \code{extracted}}
}
\description{
Get all measurement_method / timestamp combinations in the archive
}
