% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_read_table.R
\name{arrow_read_table}
\alias{arrow_read_table}
\title{Read a table from an arrow database and return as an Dataset object.}
\usage{
arrow_read_table(
  table,
  db = getOption("RRDarrow", "/Volumes/RRD.Reclassification_LEEF-2/parquet/"),
  hive_style = TRUE,
  unify_schemas = FALSE,
  format = c("parquet")
)
}
\arguments{
\item{table}{Directory of the data files}

\item{db}{fully qualified path to the folder which contains all the arrow directories, Defaul: \code{getOption("RRDarrow", "/Volumes/RRD.Reclassification_LEEF-2/parquet/")}}

\item{hive_style}{see \code{\link[arrow]{open_dataset}}}

\item{unify_schemas}{see \code{\link[arrow]{open_dataset}}}

\item{format}{see \code{\link[arrow]{open_dataset}}}
}
\value{
A Dataset R6 object. Use dplyr methods on it to query the data.
}
\description{
This function does not actually fetch the data, but returns an
Dataset object which can be further processed / filtered
using \code{dplyr}. See \link{https://db.rstudio.com/r-packages/dplyr/} for
details. To fetch the actual data, execute \code{collect()}.
This function is simply a convenience and consistency wrapper around \code{\link[arrow]{open_dataset}}
}
\seealso{
\code{\link[arrow]{open_dataset}}
}
