% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_read_toc.R
\name{arrow_read_toc}
\alias{arrow_read_toc}
\title{Read \code{toc} table from an arrow database and return as a tibble.}
\usage{
arrow_read_toc(
  db = getOption("RRDarrow", "/Volumes/RRD.Reclassification_LEEF-2/parquet/"),
  from_timestamp = 20221106,
  to_timestamp = 20230812
)
}
\arguments{
\item{db}{fully qualified path to the folder which contains all the arrow directories,
in case of LEEF called \code{parquet}. Defaul: \code{getOption("RRDarrow", "/Volumes/RRD.Reclassification_LEEF-2/parquet/")}}

\item{from_timestamp}{\code{integer}. Earliest timestamp to return. Default: \code{20210920}}

\item{to_timestamp}{\code{integer}. Latest timestamp to return. Default: \code{21000101}}
}
\value{
\code{tibble} containing the data
}
\description{
This function returns the actual data.
}
