% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_CalculateDensities.R
\name{CalculateDensities}
\alias{CalculateDensities}
\title{Re-calculate densities}
\usage{
CalculateDensities(morph, meas, extrapolation.factor, cropping.factor)
}
\arguments{
\item{meas}{}
}
\description{
Re-calculate densities
}
