% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_bemovi_archive.R
\name{classify_bemovi_archive}
\alias{classify_bemovi_archive}
\title{Classify bemovi data by using the archived data}
\usage{
classify_bemovi_archive(
  archive_dir = "/Users/rainerkrug/MountainDuck/LEEFSwift3",
  magnification = 16,
  bemovi_extract_name = "bemovi_extract.yml",
  timestamps,
  classifier_constant_name,
  classifier_increasing_name,
  output,
  mc.cores = 1
)
}
\arguments{
\item{magnification}{the magnification or the bemovi videos to be reclassified}

\item{bemovi_extract_name}{the name of the \code{.yml} containing the parameter for the nalysis}

\item{timestamps}{\code{character} vector containing the timestamps to be classified}

\item{classifier_constant_name}{the classifier for temperature treatment \strong{constant}}

\item{classifier_increasing_name}{the classifier for temperature treatment \strong{increasing}}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{mc.cores}{number of cores to be used. Defaults to 1}

\item{root_dir}{directory containing the archive. It contains the following directories:
\itemize{
\item LEEF.archived.data
\item LEEF.archived.data_segments
\item LEEF.backend.data
\item LEEF.backend.data_segments
}}
}
\value{
invisible \code{NULL}
}
\description{
Classify bemovi data by using the archived data
}
\examples{


}
