% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_bemovi_files.R
\name{classify_bemovi_files}
\alias{classify_bemovi_files}
\title{Classify algae_traits data.frame}
\usage{
classify_bemovi_files(
  datadir,
  bemovi_extract_name = "bemovi_extract.yml",
  classifier_constant_name,
  classifier_increasing_name
)
}
\arguments{
\item{datadir}{\code{character} vector containing the root directory for all files}

\item{classifier_constant_name}{\code{character} vector of name of the classifier
for temperature treatment \strong{constant} including path}

\item{classifier_increasing_name}{\code{character} vector of name of the
classifier for temperature treatment \strong{increasing} including path}

\item{algae_traits_name}{\code{character} vector containing the name of the algae
traits file without path}
}
\value{
\code{list} containing two objects:
- \code{algae_traits} including species
- \code{algae_densities} densities of the different particles identified
}
\description{
Classify algae_traits data.frame
}
\examples{


}
