% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_flowcam_archive.R
\name{classify_flowcam_archive}
\alias{classify_flowcam_archive}
\title{Classify algae_traits data.frame}
\usage{
classify_flowcam_archive(
  archive_dir = "/Users/rainerkrug/MountainDuck/LEEFSwift3",
  timestamps,
  algae_traits_name = "algae_traits_filtered.rds",
  classifier_constant_name,
  classifier_increasing_name,
  db_path
)
}
\arguments{
\item{timestamps}{\code{character} vector containing the timestamps to be classified}

\item{classifier_constant_name}{the classifier for temperature treatment \strong{constant}}

\item{classifier_increasing_name}{the classifier for temperature treatment \strong{increasing}}

\item{db_path}{if a valid path for an existing or new database \code{LEEF.RRD.sqlite} to which the
classified data will be added.}

\item{root_dir}{directory containes the archive. It contains the following directories:
\itemize{
\item LEEF.archived.data
\item LEEF.archived.data_segments
\item LEEF.backend.data
\item LEEF.backend.data_segments
}}
}
\value{
invisible \code{NULL}
}
\description{
Classify algae_traits data.frame
}
\examples{


}
