% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_flowcam_files.R
\name{classify_flowcam_files}
\alias{classify_flowcam_files}
\title{Classify algae_traits data.frame}
\usage{
classify_flowcam_files(
  datadir,
  algae_traits_name = "algae_traits_filtered.rds",
  classifier_constant_name,
  classifier_increasing_name,
  dir = tempfile(pattern = "extracted.data_")
)
}
\arguments{
\item{datadir}{\code{character} vector containing the root directory for all files}

\item{algae_traits_name}{\code{character} vector containing the name of the algae
traits file without path}

\item{classifier_constant_name}{\code{character} vector of name of the classifier
for temperature treatment \strong{constant} including path}

\item{classifier_increasing_name}{\code{character} vector of name of the
classifier for temperature treatment \strong{increasing} including path}

\item{directrory}{for the results. Default is a temporary directory. If
another string, the directory where the results will be saved. If \code{NULL}
the objects are returned,}
}
\value{
depends on the value of \code{directory}.
\itemize{
\item \code{dir == NULL}: list` containing two objects:
\itemize{
\item \code{algae_traits} including species
\item \code{algae_densities} densities of the different particles identified
}
\item \code{dir} is a string or default:
\itemize{
\item path where the objects are stored
}
}
}
\description{
Classify algae_traits data.frame
}
\examples{


}
