% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helper.R
\name{configs}
\alias{configs}
\title{Return names of all configs in config file}
\usage{
configs(config_yml = system.file("tables.yml", package = "LEEF.analysis"))
}
\arguments{
\item{config_yml}{the config file containing the sql queries}
}
\value{
\code{ggplot} object of the plot
}
\description{
Return names of all configs in config file
}
