% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rrd_trimmed.R
\name{create_rrd_trimmed}
\alias{create_rrd_trimmed}
\title{Create trimmed down density-only version of the database}
\usage{
create_rrd_trimmed(
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  trimmed_db = gsub("\\\\.sqlite", ".trimmed.sqlite", db)
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read
from option \code{RRDdb}. If not set, defaults to option \code{RRDdb};
if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{trimmed_db}{fully qualified path to the trimmed sqlite database.
Defaults to the db with the extension \code{trimmed.sqlite} instead of \code{sqlite}.}
}
\value{
the path and name of the trimmed database
}
\description{
This function creates a trimmed down version of the database which
does not contain the tables only necessary for the classification.
It does not do anything with the original database.
}
