% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_disconnect.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect the \code{tbl_SQLiteConnection}}
\usage{
db_disconnect(x)
}
\arguments{
\item{x}{a \code{tbl_SQLiteConnection} object}
}
\value{
invisibly \code{NULL}
}
\description{
Disconnect the object returned by the \code{db_read_...} functions.
}
