% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read_light_decline.R
\name{db_read_light_decline}
\alias{db_read_light_decline}
\title{Read \code{light_decline} from RRD database and return as an \code{tbl_SQLiteConnection}.}
\usage{
db_read_light_decline(db = getOption("RRDdb", "LEEF.RRD.sqlite"))
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}
}
\value{
Link to the table \code{density} in the RRD to be used with \code{dplyr} and friends
}
\description{
This function does not actually fetch the data,
but returns an \code{tbl_SQLiteConnection} object which can be further processed / filtered
using \code{dplyr}.
See \link{https://db.rstudio.com/r-packages/dplyr/} for details.
To fetch the actual data, execute \code{collect()}.
}
