% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_flowcytometer_archive.R
\name{density_flowcytometer_archive}
\alias{density_flowcytometer_archive}
\title{Gate and extract densities from flowcytometer data by using the archived data}
\usage{
density_flowcytometer_archive(
  extracted_dir =
    "/Volumes/LEEF-1_archive/LEEF.archived.data/LEEF/3.archived.data/extracted/",
  gates_coordinates,
  timestamps,
  output,
  use_H,
  log10_all = FALSE,
  min_FSC.A = NULL,
  particles = NULL,
  mc.cores = 1
)
}
\arguments{
\item{extracted_dir}{srchive directory of the extracted data}

\item{gates_coordinates}{the \code{gates_coordinates}}

\item{timestamps}{\code{character} vector containing the timestamps to be classified}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{use_H}{if \code{TRUE}, gating will be done using \code{height}, otherwie \code{area}}

\item{log10_all}{if \code{TRUE}, all data not yet log10 transformed will be log10 transformed
("FL2-A", "FL1-H", "FL2-H", "FL3-H", "FL4-H", "FSC-H", "SSC-H") in the same way as in the pipeline.}

\item{min_FSC.A}{numeric. If \code{!NULL}, \code{FSA.A <= min_FSC.A} will be fitered out by using
a rectangular filter
\code{flowCore::rectangleGate(filterId="filter_out_0", "FSC-A" = c(min_FSC.A, +Inf))}}

\item{particles}{particle class to extract. Mainly \code{bacteria} or
\code{algae}.}

\item{mc.cores}{number of cores to be used. Defaults to 1}
}
\value{
invisible \code{NULL}
}
\description{
Gate and extract densities from flowcytometer data by using the archived data
}
