% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_timestamps.R
\name{extract_timestamps}
\alias{extract_timestamps}
\title{Extract data from table}
\usage{
extract_timestamps(
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  table = NULL,
  timestamps,
  delete_data = FALSE
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{table}{the name of the table. If \code{NULL}, a list of tables in the database \code{db} will be shown.}

\item{timestamps}{timestamps which shopuld be extracted}

\item{delete_data}{\strong{Attention!} If \code{TRUE} the data is deleted from the database!}
}
\value{
invisibly NULL
}
\description{
Extract data from \code{table} matching the \code{timestamps}.
For each \code{timestamp} the data will be saved as an \code{rds} file.
If \code{delete_data == TRUE} the data will be deleted from the database.
}
