% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_traits_flowcytometer_archive.R
\name{extract_traits_flowcytometer_archive}
\alias{extract_traits_flowcytometer_archive}
\title{Extract traits from flowcytometer data by using the archived data}
\usage{
extract_traits_flowcytometer_archive(
  extracted_dir =
    "~/Desktop/flowcytometer.FIXED/LEEF.FIXED.archived.data/LEEF/3.archived.data/extracted/",
  particles = "bacteria",
  timestamps,
  output,
  length_slope = 4.933e-06,
  length_intercept = 0.2216,
  mc.cores = 1
)
}
\arguments{
\item{particles}{particle class to extract. Mainly \code{bacteria} or
\code{algae}, See \code{LEEF.measurement.flowcytometer::extract_traits()}
for details.}

\item{timestamps}{\code{character} vector containing the timestamps to be
classified}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{length_slope}{slope of the linear regression of FSC.A and size ( lm(mean_FSC.A ~ diameter_micrometer )}

\item{length_intercept}{intercept of the linear regression of FSC.A and size ( lm(mean_FSC.A ~ diameter_micrometer )}

\item{mc.cores}{number of cores to be used. Defaults to 1}
}
\value{
invisible \code{NULL}
}
\description{
Extract traits from flowcytometer data by using the archived data
}
