% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_flowcytometer_archive.R
\name{gate_flowcytometer_archive}
\alias{gate_flowcytometer_archive}
\title{Gate flowcytometer data by using the archived data}
\usage{
gate_flowcytometer_archive(
  extracted_dir =
    "/Volumes/LEEF-1_archive/LEEF.archived.data/LEEF/3.archived.data/extracted/",
  gates_coordinates,
  timestamps,
  output,
  mc.cores = 1
)
}
\arguments{
\item{timestamps}{\code{character} vector containing the timestamps to be classified}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{mc.cores}{number of cores to be used. Defaults to 1}
}
\value{
invisible \code{NULL}
}
\description{
Gate flowcytometer data by using the archived data
}
