% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF.analysis-package.R
\docType{package}
\name{LEEF.analysis-package}
\alias{LEEF.analysis}
\alias{LEEF.analysis-package}
\title{LEEF.analysis: Access Functons, Tests and Basic Analysisof the RRD Data from the LEEF Project}
\description{
Provides simple access functions to read data out of the sqlite RRD database. SQL queries can be configured in a yaml config file and used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LEEF-UZH/LEEF.analysis}
  \item Report bugs at \url{https://github.com/LEEF-UZH/LEEF.analysis/issues}
}

}
\author{
\strong{Maintainer}: Rainer M. Krug \email{Rainer.Krug@uzh.ch}

Other contributors:
\itemize{
  \item SNF Project 310030_188431 [funder]
}

}
\keyword{internal}
