% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_biomass_bemovi.R
\name{LEEF_2_biomass_bemovi}
\alias{LEEF_2_biomass_bemovi}
\title{LEEF-1 - Add biomass to traits}
\usage{
LEEF_2_biomass_bemovi(
  ciliate_traits,
  ciliate_density,
  extrapolation_factor,
  cropping_factor
)
}
\arguments{
\item{ciliate_density}{density as read from file \code{mean_density_per_ml_TIMESTAMP.rds}}
}
\value{
list containing two objects, \code{traits} containing complete traits file
  as the argument \code{algai_traits} day includinc biomass column and \code{biomasses}
  per timestamp, bottle and species per milliliter.
}
\description{
LEEF-1 - Add biomass to traits
}
