% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_biomass_flowcam.R
\name{LEEF_2_biomass_flowcam}
\alias{LEEF_2_biomass_flowcam}
\title{LEEF-2 - Add biomass to traits}
\usage{
LEEF_2_biomass_flowcam(algae_traits, algae_density)
}
\arguments{
\item{algae_traits}{algae traits as read from file \code{algae_traits_TIMESTAMP.rds}}

\item{algae_density}{algae density as read from file \code{algae_density_TIMESTAMP.rds}}
}
\value{
list containing two objects, \code{traits} containing complete traits file
  as the argument \code{algai_traits} day includinc biomass column and \code{biomasses}
  per timestamp, bottle and species per milliliter.
}
\description{
LEEF-2 - Add biomass to traits
}
