% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2.classify_bemovi_archive.R
\name{LEEF_2_classify_bemovi_archive}
\alias{LEEF_2_classify_bemovi_archive}
\title{Classify bemovi data by using the archived data}
\usage{
LEEF_2_classify_bemovi_archive(
  extracted_dir = NULL,
  magnification = 16,
  bemovi_extract_name = NULL,
  timestamps,
  classifier = NULL,
  output,
  exclude_videos = NULL,
  mc.cores = 1
)
}
\arguments{
\item{magnification}{the magnification or the bemovi videos to be reclassified}

\item{bemovi_extract_name}{the name of the \code{.yml} containing the parameter for
the analysis in the directory of the data. The directory depends on the
\code{root_dir}, \code{magnification} and \code{timestamp}}

\item{timestamps}{\code{character} vector containing the timestamps to be classified}

\item{classifier}{classifier for bemovi 16 including path}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{exclude_videos}{file names of videos to exclude. If \code{NULL}, all will be used.}

\item{mc.cores}{number of cores to be used. Defaults to 1}

\item{directory}{with extracted data}
}
\value{
invisible \code{NULL}
}
\description{
Classify bemovi data by using the archived data
}
