% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_classify_bemovi_files.R
\name{LEEF_2_classify_bemovi_files}
\alias{LEEF_2_classify_bemovi_files}
\title{Classify algae_traits data.frame}
\usage{
LEEF_2_classify_bemovi_files(
  datadir,
  bemovi_extract_name = NULL,
  classifier = NULL,
  exclude_videos = NULL
)
}
\arguments{
\item{datadir}{\code{character} vector containing the root directory for all files}

\item{bemovi_extract_name}{\code{character} vector containing the name of the
bemovi config file including path.}

\item{classifier}{classifier for bemovi 16 including path}

\item{exclude_videos}{file names of videos to exclude. If \code{NULL}, all will be used.}
}
\value{
\code{list} containing three objects:
- \code{mean_density_per_ml}
- \code{morph_mvt}
- \code{trajectory_data}
}
\description{
Classify algae_traits data.frame
}
