% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_classify_flowcam_archive.R
\name{LEEF_2_classify_flowcam_archive}
\alias{LEEF_2_classify_flowcam_archive}
\title{Classify algae_traits data.frame}
\usage{
LEEF_2_classify_flowcam_archive(
  extracted_dir = NULL,
  timestamps,
  algae_traits_name = "algae_traits_filtered.rds",
  classifier = NULL,
  species_tracked = NULL,
  output,
  mc.cores = 1,
  bottle = NULL
)
}
\arguments{
\item{extracted_dir}{path to the extracted data}

\item{timestamps}{\code{character} vector containing the timestamps to be classified}

\item{classifier}{classifier for flowcytometer}

\item{species_tracked}{names of the species tracked as a character vector. If \code{NULL} it will be read from the original configuration file in the \code{datadir}.}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{mc.cores}{number of cores to be used. Defaults to 1}

\item{bottle}{if not 'NULL' (default) only classify this bottle. Needs to be a single bottle!}
}
\value{
invisible \code{NULL}
}
\description{
Classify algae_traits data.frame
}
