% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_classify_flowcam_files.R
\name{LEEF_2_classify_flowcam_files}
\alias{LEEF_2_classify_flowcam_files}
\title{Classify algae_traits data.frame}
\usage{
LEEF_2_classify_flowcam_files(
  datadir,
  algae_traits_name = "algae_traits_filtered.rds",
  classifier = NULL,
  timestamp = "55555555",
  species_tracked = NULL,
  bottle = NULL
)
}
\arguments{
\item{datadir}{\code{character} vector containing the root directory for all files}

\item{algae_traits_name}{\code{character} vector containing the name of the algae
traits file without path}

\item{classifier}{classifier for flowcam}

\item{timestamp}{timestamp to be used to stamp the classified data}

\item{species_tracked}{names of the species tracked as a character vector.
If \code{NULL} it will be read from the original configuration file in the \code{datadir}.}

\item{bottle}{if not 'NULL' (default) only classify this bottle. Needs to be a single bottle!}
}
\value{
\code{list} containing two objects:
- \code{algae_traits} including species
- \code{algae_densities} densities of the different particles identified
}
\description{
Classify algae_traits data.frame
}
