% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_density_flowcytometer_archive.R
\name{LEEF_2_density_flowcytometer_archive}
\alias{LEEF_2_density_flowcytometer_archive}
\title{Gate and extract densities from flowcytometer data by using the archived data}
\usage{
LEEF_2_density_flowcytometer_archive(
  extracted_dir,
  timestamps,
  output,
  use_H,
  gates_coordinates,
  min_FSC.A = NULL,
  particles = NULL,
  mc.cores = 5
)
}
\arguments{
\item{extracted_dir}{srchive directory of the extracted data}

\item{timestamps}{\code{character} vector containing the timestamps to be classified}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{use_H}{if \code{TRUE}, gating will be done using \code{height}, otherwie \code{area}}

\item{gates_coordinates}{the \code{gates_coordinates}}

\item{min_FSC.A}{numeric. If \code{!NULL}, \code{FSA.A <= min_FSC.A} will be fitered out by using
a rectangular filter
\code{flowCore::rectangleGate(filterId="filter_out_0", "FSC-A" = c(min_FSC.A, +Inf))}}

\item{particles}{particle class to extract. Mainly \code{bacteria} or
\code{algae}.}
}
\value{
invisible \code{NULL}
}
\description{
Gate and extract densities from flowcytometer data by using the archived data
}
