% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_extract_traits_flowcytometer_archive.R
\name{LEEF_2_extract_traits_flowcytometer_archive}
\alias{LEEF_2_extract_traits_flowcytometer_archive}
\title{Extract traits from flowcytometer data by using the archived data}
\usage{
LEEF_2_extract_traits_flowcytometer_archive(
  extracted_dir =
    "~/Desktop/flowcytometer.FIXED/LEEF.FIXED.archived.data/LEEF/3.archived.data/extracted/",
  gates_coordinates,
  particles = "bacteria",
  timestamps,
  output,
  length_slope,
  length_intercept,
  use_H,
  min_FSC.A,
  log10_all = FALSE,
  wellid_keyword = "$WELLID"
)
}
\arguments{
\item{gates_coordinates}{the \code{gates_coordinates}}

\item{particles}{particle class to extract. Mainly \code{bacteria} or
\code{algae}, See \code{LEEF.measurement.flowcytometer::extract_traits()}
for details.}

\item{timestamps}{\code{character} vector containing the timestamps to be
classified}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{length_slope}{slope of the linear regression of FSC.A and size ( lm(mean_FSC.A ~ diameter_micrometer )}

\item{length_intercept}{intercept of the linear regression of FSC.A and size ( lm(mean_FSC.A ~ diameter_micrometer )}

\item{use_H}{if \code{TRUE}, gating will be done using \code{height}, otherwie \code{area}}

\item{min_FSC.A}{numeric. If \code{!NULL}, \code{FSA.A <= min_FSC.A} will be fitered out by using
a rectangular filter
\code{flowCore::rectangleGate(filterId="filter_out_0", "FSC-A" = c(min_FSC.A, +Inf))}}

\item{log10_all}{if \code{TRUE}, all data not yet log10 transformed will be log10 transformed
("FL2-A", "FL1-H", "FL2-H", "FL3-H", "FL4-H", "FSC-H", "SSC-H") in the same way as in the pipeline.}

\item{wellid_keyword}{the kwyword which is used to identify the well ID. Usually "$WELLID" (default), but for the EAWAG Flowcytometer it is "$SMNO".}

\item{mc.cores}{number of cores to be used. Defaults to 1}
}
\value{
invisible \code{NULL}
}
\description{
Extract traits from flowcytometer data by using the archived data
}
