% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LEEF_2_plot_conductivity_per_bottle_per_timestamp.R
\name{LEEF_2_plot_conductivity_per_bottle_per_timestamp}
\alias{LEEF_2_plot_conductivity_per_bottle_per_timestamp}
\title{Plot diagnostic plot to check number of species per timestamp}
\usage{
LEEF_2_plot_conductivity_per_bottle_per_timestamp(
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  treatment_begin_day = 70,
  treatment_end_day = 154
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{treatment_begin_day}{begin of treatment (vertical red line in plot). If \code{NULL} none is plotted.}

\item{treatment_end_day}{end of treatment (vertical red line in plot). If \code{NULL} none is plotted.}
}
\value{
\code{ggplot} object of the plot
}
\description{
Plot diagnostic plot to check number of species per timestamp
}
