% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LEEF_2_plot_flowcytometer_individual_per_bottle.R
\name{LEEF_2_plot_flowcytometer_individual_per_bottle_per_timestamp}
\alias{LEEF_2_plot_flowcytometer_individual_per_bottle_per_timestamp}
\title{Plot diagnostic plot to check is plte has an impact on flowcytometer for LEEF-2}
\usage{
LEEF_2_plot_flowcytometer_individual_per_bottle_per_timestamp(
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  species = c("bacteria")
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{species}{species to be plotted}
}
\value{
\code{ggplot} object of the plot
}
\description{
Plot diagnostic plot to check is plte has an impact on flowcytometer for LEEF-2
}
