% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_renamer.R
\name{LEEF_2_rename_composition}
\alias{LEEF_2_rename_composition}
\title{FUNCTION_TITLE}
\usage{
LEEF_2_rename_composition(object)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
