% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_renamer.R
\name{LEEF_2_rename_species}
\alias{LEEF_2_rename_species}
\title{FUNCTION_TITLE}
\usage{
LEEF_2_rename_species(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
