% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2_report_diagnostic.R
\name{LEEF_2_report_diagnostic}
\alias{LEEF_2_report_diagnostic}
\title{Report for checking pipeline and sampling}
\usage{
LEEF_2_report_diagnostic(db, suffix = "", format = "html", lastDays = 7)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. The report will be saved to the same directory.}

\item{suffix}{suffix for the file name}

\item{format}{the format of the report as a character vector of length 1.
Supported are at the moment: \code{html} (the default), \code{pdf} and \code{word}.}

\item{lastDays}{number of last days to be included in graph 2. Default: 7}
}
\value{
the fully qualified file name to the created report.
}
\description{
Report for checking pipeline and sampling
}
