% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_view_o2.R
\name{make_view_o2}
\alias{make_view_o2}
\title{Create \code{o2} view wgich contains selected fields from the os measurement}
\usage{
make_view_o2(db = getOption("RRDdb", "LEEF.RRD.sqlite"), overwrite = FALSE)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{overwrite}{if \code{TRUE}, overwrite existing view}
}
\value{
the result of the execution of the ecreatiuon of the view.
}
\description{
Create \code{o2} view wgich contains selected fields from the os measurement
}
