% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_views.R
\name{make_views}
\alias{make_views}
\title{Create all views}
\usage{
make_views(
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  start_date = "2021-09-20",
  overwrite = FALSE
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{start_date}{the start date of the experiment}

\item{overwrite}{if \code{TRUE}, overwrite existing view}
}
\value{
invisibly \code{NULL}
}
\description{
This function calls the other \code{make_view_...()} functions, wrapped in individual \code{try()} blocks.
}
