% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{new_parquet_definition}
\alias{new_parquet_definition}
\title{FUNCTION_TITLE}
\usage{
new_parquet_definition(sources = NULL, partitioning = NULL, hive_style = TRUE)
}
\arguments{
\item{sources}{PARAM_DESCRIPTION, Default: NULL}

\item{partitioning}{PARAM_DESCRIPTION, Default: NULL}

\item{hive_style}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
}
\seealso{
\code{\link[arrow]{open_dataset}}
}
