% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlays.R
\name{overlays}
\alias{overlays}
\title{Create}
\usage{
overlays(
  timestamp = 20210920,
  magnification = 25,
  cropped = FALSE,
  label = "species",
  overlay.type = "both",
  crf = 17,
  gamma = 2,
  ffmpeg = "ffmpeg",
 
    from_current_to_archive_dir = "./../../../../../Duck/LEEFSwift3/LEEF.archived.data/LEEF/3.archived.data",
  mc.cores = 7
)
}
\arguments{
\item{timestamp}{the timestamp of the sampling day}

\item{magnification}{the magnification of the bemovi method}

\item{cropped}{if \code{TRUE}, the cropped measurement should be used}

\item{label}{column to be used to label the particle. Default
\code{"pecies"}.}

\item{overlay.type}{Overlays can either be shown as "label", "circle" or
"both". Default: \code{"both"}}

\item{crf}{integer value between 1 to 51, where 1 means lossless, 17 is nearly visually lossless,
51 is worst quality. Default value is 23}

\item{gamma}{increase video dynamic range. Value between 0.1 and 10. Default 2. see \url{https://ffmpeg.org/ffmpeg-filters.html#eq} for further info}

\item{ffmpeg}{command to run ffmpeg. It can include a path. Default
\code{ffmpeg}}

\item{from_current_to_archive_dir}{path from the current directory to the
archived data can be found and the folders \code{extracted} and
\code{pre-processed}}

\item{mc.cores}{number of cores to be used for parallel execution. Defaults
to 1}
}
\value{

}
\description{
Create
}
