% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlays_from_folders.R
\name{overlays_from_folders}
\alias{overlays_from_folders}
\title{Create overlays}
\usage{
overlays_from_folders(
  traj_data_file,
  avi_folder,
  bemovi_extract_yml_file,
  temp_overlay_folder,
  overlay_folder,
  overlay_type = "both",
  label = "species",
  ffmpeg = "ffmpeg",
  font_size = 24,
  circle_size = 120,
  crf = 23,
  gamma = 2,
  mc_cores = 1,
  overwrite = FALSE
)
}
\arguments{
\item{traj_data_file}{file name of the file containing the trajectory data (usually from the Master file)}

\item{avi_folder}{folder containing the input \code{.avi} files}

\item{bemovi_extract_yml_file}{name of the \code{bemovi_extract.yml} config file}

\item{temp_overlay_folder}{directory where the subtitle files will be saved}

\item{overlay_folder}{directory where the overlay video will be saved}

\item{overlay_type}{option for the overlays. Overlays can either be shown as
\code{"label"}, \code{"circle"} or \code{"both"}}

\item{label}{column to be used to label the particle. Default is
\code{"trajectory"}, other useful might be \code{"species"}}

\item{ffmpeg}{command to run ffmpeg. The default is \code{par_ffmpeg()}. It
can include a path.}

\item{font_size}{size of the font for the labels. Default: 24}

\item{circle_size}{size of the circle. Default: 120}

\item{crf}{integer value between 1 to 51, where 1 means lossless, 17 is
nearly visually lossless, 51 is worst quality. Default value is 23}

\item{gamma}{gamma correction. Value between 0.1 and 10. Default 2. see
\url{https://ffmpeg.org/ffmpeg-filters.html#eq} for further info}

\item{mc_cores}{number of cores toi be used for parallel execution.
Defaults to \code{par_mc.cores()}}

\item{overwrite}{if \code{TRUE}, existing overlays will be overwritten. Default is \code{FALSE}}
}
\description{
This is a wrapper around the function \code{bemovi.LEEF::create_overlays_subtitle_directory()}
providing default values for LEEF.
Overlays will be created from a folder containing the video files and the bemovi config file
and the trajectory data file name.
}
