% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{parquet_add_bemovi_16}
\alias{parquet_add_bemovi_16}
\alias{parquet_add_bemovi_25}
\alias{parquet_add_bemovi_25_cropped}
\alias{parquet_add_flowcam}
\alias{parquet_add_flowcytometer}
\alias{parquet_add_o2}
\alias{parquet_add_manualcount}
\alias{parquet_add_toc}
\alias{parquet_add_conductivity}
\alias{object_to_parquet}
\title{FUNCTION_TITLE}
\usage{
parquet_add_bemovi_16(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy",
  rename = FALSE
)

parquet_add_bemovi_25(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy",
  rename = FALSE
)

parquet_add_bemovi_25_cropped(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy",
  rename = FALSE
)

parquet_add_flowcam(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy",
  rename = FALSE
)

parquet_add_flowcytometer(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy",
  rename = FALSE
)

parquet_add_o2(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy",
  rename = FALSE
)

parquet_add_manualcount(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy",
  rename = FALSE
)

parquet_add_toc(
  fn = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy",
  rename = FALSE
)

parquet_add_conductivity(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy",
  rename = FALSE
)

object_to_parquet(
  object,
  path_to_parquet,
  partition = "no",
  compression = "snappy",
  compression_level = NULL,
  ...
)
}
\arguments{
\item{fns}{PARAM_DESCRIPTION, Default: NULL}

\item{path_to_parquet_root_dir}{PARAM_DESCRIPTION, Default: NULL}

\item{compression}{compression as used in \code{\link[arrow]{write_parquet}}, Default: "snappy"}

\item{rename}{Should data be renamed to follow specified standards, Default: FALSE}

\item{fn}{PARAM_DESCRIPTION, Default: NULL}

\item{object}{object to be written as a parquet file / arrow db}

\item{path_to_parquet}{PARAM_DESCRIPTION, Default: NULL}

\item{compression_level}{PARAM_DESCRIPTION, Default: "NULL"}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

Essentially combined re-implementation of \code{\link[parquetize]{rds_to_parquet}} 
and \code{\link[parquetize]{csv_to_parquet}} plus standardisation on small 
letter column names and timestamps as character.
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
}
\seealso{
\code{\link[pbapply]{pbapply}}

\code{\link[pbapply]{pbapply}}

\code{\link[pbapply]{pbapply}}

\code{\link[pbapply]{pbapply}}

\code{\link[pbapply]{pbapply}}

\code{\link[pbapply]{pbapply}}

\code{\link[pbapply]{pbapply}}
}
