% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-parquet.R, R/parquet.R
\name{parquet_add_manualcount}
\alias{parquet_add_manualcount}
\title{FUNCTION_TITLE}
\usage{
parquet_add_manualcount(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy"
)

parquet_add_manualcount(
  fns = NULL,
  path_to_parquet_root_dir = NULL,
  compression = "snappy"
)
}
\arguments{
\item{fns}{PARAM_DESCRIPTION, Default: NULL}

\item{path_to_parquet_root_dir}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
}
\seealso{
\code{\link[pbapply]{pbapply}}
 \code{\link[parquetize]{csv_to_parquet}}

\code{\link[pbapply]{pbapply}}
 \code{\link[parquetize]{csv_to_parquet}}
}
