% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet_read_table.R
\name{parquet_open_table}
\alias{parquet_open_table}
\title{Open Arrow DB with parquet files}
\usage{
parquet_open_table(def)
}
\arguments{
\item{def}{Path to the parquet files definition or the actual definition object. These have usually the name `TABLENAME.sources.rds`}
}
\value{
Object to the table \code{density} in the RRD to be used with \code{dplyr} and friends
}
\description{
Opens the arrow dataset of parquet files. The resulting pbject can be used in a dplyr pipeline as a starting point.
}
