% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_o2s_per_bottle_per_timestamp.R
\name{plot_o2s_per_bottle_per_timestamp}
\alias{plot_o2s_per_bottle_per_timestamp}
\title{Plot diagnostic plot to check number of species per timestamp}
\usage{
plot_o2s_per_bottle_per_timestamp(db = getOption("RRDdb", "LEEF.RRD.sqlite"))
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}
}
\value{
\code{ggplot} object of the plot
}
\description{
Plot diagnostic plot to check number of species per timestamp
}
