% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rrd.R
\name{prepare_rrd}
\alias{prepare_rrd}
\title{Prepare the RRD database for usage with the other functions in this package.}
\usage{
prepare_rrd(
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  composition = NULL,
  experimetal_design = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{composition}{fully qualified path to the \code{csv} file with the composition data}

\item{experimetal_design}{fully qualified path to the \code{csv} file with the experimetal_design data}

\item{overwrite}{if \code{TRUE}, the existing tables / views will be overwritten.}
}
\description{
This function adds the needed tables and creates the vuews needed for the analysis.
}
