% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Read data from the RRD database}
\usage{
read_sql(db = getOption("RRDdb", "LEEF.RRD.sqlite"), sql)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to \code{LEEF.RRD.sqlite}}

\item{sql}{sql statement}
}
\value{
the table resulting from the query as a \code{data.frame} object.
}
\description{
The function executes the provided sql statement and returns the resulting table0.
For security, the database is opened in \bold{read-only} mode.!
}
