% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_diagnostic.R
\name{report_diagnostic}
\alias{report_diagnostic}
\title{Report for checking pipeline and sampling}
\usage{
report_diagnostic(db)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. The report will be saved to the same directory.}
}
\value{
the fully qualified file name to the created report.
}
\description{
Report for checking pipeline and sampling
}
