% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_interactive.R
\name{report_interactive}
\alias{report_interactive}
\title{Run interactive report}
\usage{
report_interactive(db)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. The report will be saved to the same directory.}
}
\value{
invisibkly \code{NULL}
}
\description{
Run interactive report
}
