% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRD_create_indices.R
\name{RRD_create_indices}
\alias{RRD_create_indices}
\title{Create indices in RRD database}
\usage{
RRD_create_indices(dbname, LEEF = NULL)
}
\arguments{
\item{dbname}{the path and name of the database. Must exist!}

\item{either}{"LEEF-1" or "LEEF-2"}
}
\description{
Create indices in RRD database
}
