% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRD_new.R
\name{RRD_new}
\alias{RRD_new}
\title{Create new RRD database}
\usage{
RRD_new(dbname, LEEF = NULL)
}
\arguments{
\item{dbname}{the path and name of the database. Must not exist.}

\item{either}{"LEEF-1" or "LEEF-2"}
}
\description{
Create a new database following the scheme used for the LEEF.RRD database
}
