% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitychecks.R
\name{sanity_check}
\alias{sanity_check}
\title{Sanity check data for pipeline}
\usage{
sanity_check(sample_dir = ".")
}
\arguments{
\item{sample_dir}{root dir in which the folder \code{00.general.parameter} and \code{0.raw.data}
are located}
}
\value{
named \code{list} object, with one element per method. If the test passed
  for the method, the result is \code{TRUE}, if it fails, a string indicating the
  reason why it failed
}
\description{
Executes onle test per method required.
}
