% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitychecks.R
\name{sanity_check_bemovi.mag.16}
\alias{sanity_check_bemovi.mag.16}
\title{Sanity check for bemovi.mag.16}
\usage{
sanity_check_bemovi.mag.16(sample_dir = ".")
}
\arguments{
\item{sample_dir}{root dir in which the folder \code{bemovi.mag.16} and \code{0.raw.data}
are located}
}
\value{
\code{TRUE} if the test passed, a string indicating the error if it failed.
}
\description{
Sanity check for bemovi.mag.16
}
