% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitychecks.R
\name{sanity_check_flowcam}
\alias{sanity_check_flowcam}
\title{Sanity check for flowcam}
\usage{
sanity_check_flowcam(sample_dir = ".")
}
\arguments{
\item{sample_dir}{root dir in which the folder \code{flowcam} and \code{0.raw.data}
are located}
}
\value{
\code{TRUE} if the test passed, a string indicating the error if it failed.
}
\description{
Sanity check for flowcam
}
