% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_get_timestamp.R
\name{sanity_check_manualcount}
\alias{sanity_check_manualcount}
\title{Sanity check for manualcount}
\usage{
sanity_check_manualcount(sample_dir = ".")
}
\arguments{
\item{sample_dir}{root dir in which the folder\code{00.general.parameter} and \code{0.raw.data}
are located}
}
\value{
\code{TRUE} if the test passed, a string indicating the error if it failed.
}
\description{
Sanity check for manualcount
}
