% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitychecks.R
\name{sanity_check_o2meter}
\alias{sanity_check_o2meter}
\title{Sanity check for o2meter}
\usage{
sanity_check_o2meter(sample_dir = ".")
}
\arguments{
\item{sample_dir}{root dir in which the folder \code{00.general.parameter} and \code{0.raw.data}
are located}
}
\value{
\code{TRUE} if the test passed, a string indicating the error if it failed.
}
\description{
Sanity check for o2meter
}
