% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD_SetNotFoundSpeciesTo0.R
\name{SetNotFoundSpeciesTo0}
\alias{SetNotFoundSpeciesTo0}
\title{Set number of Not Found species to 0}
\usage{
SetNotFoundSpeciesTo0(
  densities,
  mb,
  meas,
  exp_design,
  compositions,
  species.tracked
)
}
\arguments{
\item{species.tracked}{}
}
\description{
ADD DESCRIPTION
}
