% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{species_set}
\alias{species_set}
\title{Defines sets of species}
\usage{
species_set(species_set_id = NULL)
}
\arguments{
\item{species_set_id}{if \code{NULL}, return the names of the species sets defined.
If the name or the index of a specific set, return that set.}
}
\value{
Names of sets or the set itself
}
\description{
Defines sets of species
}
