% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_overlays.R
\name{update_overlays}
\alias{update_overlays}
\title{Just a function to update the overlays.}
\usage{
update_overlays(
  params = list(cores = 7, pre_processed_folder =
    "/Volumes/LEEF-1_archive/LEEF.archived.data/LEEF/3.archived.data/pre_processed/",
    extracted_dir =
    "/Volumes/LEEF-1_archive/LEEF.archived.data/LEEF/3.archived.data/extracted/",
    output_dir = "./overlays/"),
  overwrite = FALSE
)
}
\arguments{
\item{params}{list of parameter}

\item{overwrite}{if `TRUE`, existing overlays will be ov erwritten. Default is `FALSE`}
}
\description{
Just a function to update the overlays.
}
