% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_db.R
\name{add_to_db}
\alias{add_to_db}
\title{Add data to RRD database into existing table}
\usage{
add_to_db(
  fns,
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  tables,
  remove_timestamps = NULL
)
}
\arguments{
\item{fns}{vector of \code{rds} or \code{csv} files names containing data to be added.
The data has to contain a colum named \code{timestamp}.}

\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{tables}{\code{vector} of the table names the data of the \code{fns} of
the same index should be added to.
\strong{The table has to exist and contain a field named \code{timestamp}}}

\item{remove_timestamps}{vector of timestamps to be removed.}
}
\value{
vector of length of \code{fns} with \code{TRUE} if the data has been added,
\code{FALSE} otherwise
}
\description{
Add data to RRD database into existing table
}
