% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_bemovi_files.R,
%   R/gate_flowcytometer_files.R
\name{classify_bemovi_files}
\alias{classify_bemovi_files}
\title{Classify algae_traits data.frame}
\usage{
classify_bemovi_files(
  datadir,
  bemovi_extract_name = NULL,
  classifier_constant_name,
  classifier_increasing_name
)

classify_bemovi_files(
  datadir,
  bemovi_extract_name = NULL,
  classifier_constant_name,
  classifier_increasing_name
)
}
\arguments{
\item{datadir}{\code{character} vector containing the root directory for all files}

\item{bemovi_extract_name}{\code{character} vector containing the name of the
bemovi config file including path.}

\item{classifier_constant_name}{\code{character} vector of name of the classifier
for temperature treatment \strong{constant} including path}

\item{classifier_increasing_name}{\code{character} vector of name of the
classifier for temperature treatment \strong{increasing} including path}
}
\value{
\code{list} containing three objects:
- \code{mean_density_per_ml}
- \code{morph_mvt}
- \code{trajectory_data}

\code{list} containing three objects:
- \code{mean_density_per_ml}
- \code{morph_mvt}
- \code{trajectory_data}
}
\description{
Classify algae_traits data.frame

Classify algae_traits data.frame
}
