% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_flowcam_archive.R
\name{classify_flowcam_archive}
\alias{classify_flowcam_archive}
\title{Classify algae_traits data.frame}
\usage{
classify_flowcam_archive(
  archive_dir = "/Users/rainerkrug/MountainDuck/LEEFSwift3",
  timestamps,
  algae_traits_name = "algae_traits_filtered.rds",
  classifier_constant_name,
  classifier_increasing_name,
  species_tracked = NULL,
  output,
  mc.cores = 1
)
}
\arguments{
\item{timestamps}{\code{character} vector containing the timestamps to be classified}

\item{classifier_constant_name}{the classifier for temperature treatment \strong{constant}}

\item{classifier_increasing_name}{the classifier for temperature treatment \strong{increasing}}

\item{species_tracked}{names of the species tracked as a character vector.
If \code{NULL} it will be read from the original configuration file in the \code{datadir}.}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{mc.cores}{number of cores to be used. Defaults to 1}

\item{root_dir}{directory containes the archive. It contains the following directories:
\itemize{
\item LEEF.archived.data
\item LEEF.archived.data_segments
\item LEEF.backend.data
\item LEEF.backend.data_segments
}}
}
\value{
invisible \code{NULL}
}
\description{
Classify algae_traits data.frame
}
