% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_traits_flowcytometer_archive.R
\name{extract_traits_flowcytometer_archive}
\alias{extract_traits_flowcytometer_archive}
\title{Extract traits from flowcytometer data by using the archived data}
\usage{
extract_traits_flowcytometer_archive(
  extracted_dir =
    "/Volumes/LEEF-1_archive/LEEF.archived.data/LEEF/3.archived.data/extracted/",
  particles = "bacteria",
  timestamps,
  output,
  mc.cores = 1
)
}
\arguments{
\item{particles}{particle class to extract. Mainly \code{bacteria} or
\code{algae}, See \code{LEEF.measurement.flowcytometer::extract_traits()}
for details.}

\item{timestamps}{\code{character} vector containing the timestamps to be
classified}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{mc.cores}{number of cores to be used. Defaults to 1}
}
\value{
invisible \code{NULL}
}
\description{
Extract traits from flowcytometer data by using the archived data
}
