% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEEF_2.classify_bemovi_archive.R
\name{LEEF_2_classify_bemovi_archive}
\alias{LEEF_2_classify_bemovi_archive}
\title{Classify bemovi data by using the archived data}
\usage{
LEEF_2_classify_bemovi_archive(
  archive_dir = "/Users/rainerkrug/MountainDuck/LEEFSwift3",
  magnification = 16,
  bemovi_extract_name = "bemovi_extract.yml",
  species_tracked = NULL,
  timestamps,
  classifier_name = "svm_video_classifiers_16x.rds",
  output,
  exclude_videos = NULL,
  mc.cores = 1
)
}
\arguments{
\item{magnification}{the magnification or the bemovi videos to be reclassified}

\item{bemovi_extract_name}{the name of the \code{.yml} containing the parameter for
the analysis in the directory of the data. The directory depends on the
\code{root_dir}, \code{magnification} and \code{timestamp}}

\item{species_tracked}{names of the species tracked as a character vector. If
\code{NULL} the species tracked from the \code{bemovi_extract_file}  will be used.}

\item{timestamps}{\code{character} vector containing the timestamps to be classified}

\item{classifier_name}{\code{character} vector of name of the classifier for bemovi 16 including path}

\item{output}{path to which the classified data will be saved as \code{rds}}

\item{exclude_videos}{file names of videos to exclude. If \code{NULL}, all will be used.}

\item{mc.cores}{number of cores to be used. Defaults to 1}
}
\value{
invisible \code{NULL}
}
\description{
Classify bemovi data by using the archived data
}
