% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlays.R
\name{overlays}
\alias{overlays}
\title{Create}
\usage{
overlays(
  timestamp = 20210920,
  magnification = 25,
  cropped = FALSE,
  label = "species",
  overlay.type = "both",
  hq = FALSE,
  ffmpeg = "ffmpeg",
 
    from_current_to_archive_dir = "./../../../../../Duck/LEEFSwift3/LEEF.archived.data/LEEF/3.archived.data",
  mc.cores = 7
)
}
\arguments{
\item{timestamp}{the timestamp of the sampling day}

\item{magnification}{the magnification of the bemovi method}

\item{cropped}{if \code{TRUE}, the cropped measurement should be used}

\item{label}{column to be used to label the particle. Default
\code{"pecies"}.}

\item{overlay.type}{Overlays can either be shown as "label", "circle" or
"both". Default: \code{"both"}}

\item{hq}{if \code{TRUE}, overlays are in same quality as the original video,
if \code{FALSE} they will be downsampled. Default \code{FALSE}}

\item{ffmpeg}{command to run ffmpeg. It can include a path. Default
\code{ffmpeg}}

\item{from_current_to_archive_dir}{path from the current directory to the
archived data can be found and the folders \code{extracted} and
\code{pre-processed}}

\item{mc.cores}{number of cores to be used for parallel execution. Defaults
to 1}
}
\value{

}
\description{
Create
}
