% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlays_from_reclassification.R
\name{overlays_from_reclassification}
\alias{overlays_from_reclassification}
\title{Create overlays}
\usage{
overlays_from_reclassification(
  archive_dir = "/Users/rainerkrug/MountainDuck/LEEFSwift3",
  classification_dir = "~/RRD.Reclassification_1",
  magnification = 16,
  yml_postfix = NULL,
  timestamp,
  overlay_folder = NULL,
  overlay_type = "both",
  label = "species",
  ffmpeg = "ffmpeg",
  font_size = 24,
  circle_size = 120,
  crf = 23,
  gamma = 2,
  mc_cores = 1
)
}
\arguments{
\item{magnification}{the magnification of the bemovi method}

\item{timestamp}{the timestamp of the sampling day}

\item{label}{column to be used to label the particle. Default
\code{"pecies"}.}

\item{ffmpeg}{command to run ffmpeg. It can include a path. Default
\code{ffmpeg}}

\item{crf}{integer value between 1 to 51, where 1 means lossless, 17 is nearly visually lossless,
51 is worst quality. Default value is 23}

\item{gamma}{increase video dynamic range. Value between 0.1 and 10. Default 2. see \url{https://ffmpeg.org/ffmpeg-filters.html#eq} for further info}

\item{cropped}{if \code{TRUE}, the cropped measurement should be used}

\item{overlay.type}{Overlays can either be shown as "label", "circle" or
"both". Default: \code{"both"}}

\item{mc.cores}{number of cores to be used for parallel execution. Defaults
to 1}

\item{from_current_to_archive_dir}{path from the current directory to the
archived data can be found and the folders \code{extracted} and
\code{pre-processed}}
}
\value{

}
\description{
Create overlays
}
