% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_per_timestamp_per_measure.R
\name{plot_species_per_timestamp_per_measure}
\alias{plot_species_per_timestamp_per_measure}
\title{Plot diagnostic plot to check species per timestamp per method}
\usage{
plot_species_per_timestamp_per_measure(
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  config_yml = system.file("tables.yml", package = "LEEF.analysis"),
  config
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. Default, read from option \code{RRDdb}.
If not set, defaults to option \code{RRDdb}; if this is not set, defaults to \code{LEEF.RRD.sqlite}}

\item{config_yml}{the config file containing an sql query named \code{species_per_bottle}}

\item{config}{the configuration in \code{list} format containing an sql query named \code{species_per_bottle}.
If not specified, the \code{config_yml} will be used}
}
\value{
\code{ggplot} object of the plot
}
\description{
Plot diagnostic plot to check species per timestamp per method
}
