% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rrd.R
\name{read_rrd}
\alias{read_rrd}
\title{Read data from the RRD database}
\usage{
read_rrd(
  db = getOption("RRDdb", "LEEF.RRD.sqlite"),
  measurements = c("bemovi_mag_16", "bemovi_mag_25", "bemovi_mag_25_cropped",
    "flowcam", "flowcytometer", "manualcount", "o2meter"),
  config_yml = system.file("tables.yml", package = "LEEF.analysis"),
  config
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database}

\item{measurements}{the configuration as int the config file specified to be
used}

\item{config_yml}{the config file containing the sql queries}

\item{config}{the configuration in \code{list} format. If not specified,
the \code{config_yml} will be used}
}
\value{
if \code{length(measurements) > 1} a list containing the returned tables, otherwise the table as a
  \code{data.frame} object.
}
\description{
The function uses a config file in yaml format to either build an sql query
or the sql query provided to execute the query and return the data in
data.frame format. In the config file, either an \code{SQL} statement or a
\code{FROM} and \code{SELSECT} fields can be specified. If an \code{SQL}
statement is provided, it takes precedence over the others. If no \code{SQL}
is specified, the SQL statement is build from the \code{SELECT} and
\code{FROM} parts. For security, the database is opened in \bold{read-only}
mode.!
}
