% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_quarto_temp.R
\name{render_quarto_temp}
\alias{render_quarto_temp}
\title{Render a quarto file in a temporary directory and copy the resulting files back to the working directory}
\usage{
render_quarto_temp(input, output_format = "html", ...)
}
\arguments{
\item{input}{the input qmd file}

\item{output_format}{the output format}

\item{...}{additional arguments for \code{quarto::quarto_render()}}
}
\description{
Render a quarto file in a temporary directory and copy the resulting files back to the working directory
}
