% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_diagnostic.R
\name{report_diagnostic}
\alias{report_diagnostic}
\title{Report for checking pipeline and sampling}
\usage{
report_diagnostic(db, suffix = "", format = "html")
}
\arguments{
\item{db}{fully qualified path to the sqlite database. The report will be saved to the same directory.}

\item{suffix}{suffix for the file name}

\item{format}{the format of the report as a character vector of length 1.
Supported are at the moment: \code{html} (the default), \code{pdf} and \code{word}.}
}
\value{
the fully qualified file name to the created report.
}
\description{
Report for checking pipeline and sampling
}
