% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_diagnostic.R
\name{report_diagnostic}
\alias{report_diagnostic}
\title{Report for checking pipeline and sampling}
\usage{
report_diagnostic(
  db,
  template = "LEEF_1",
  suffix = "",
  format = "html",
  lastDays = 7
)
}
\arguments{
\item{db}{fully qualified path to the sqlite database. The report will be saved to the same directory.}

\item{template}{Template to be used for report. At the moment only \code{"LEEF_1} and \code{"LEEF_2} supported.}

\item{suffix}{suffix for the file name}

\item{format}{the format of the report as a character vector of length 1.
Supported are at the moment: \code{html} (the default), \code{pdf} and \code{word}.}

\item{lastDays}{number of last days to be included in graph 2. Default: 7}
}
\value{
the fully qualified file name to the created report.
}
\description{
Report for checking pipeline and sampling
}
