% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_flowcytometer.R
\name{report_flowcytometer}
\alias{report_flowcytometer}
\title{Report for checking pipeline and sampling}
\usage{
report_flowcytometer(
  timestamp = "20230106",
  extracted_base_dir =
    "~/Duck/LEEFSwift3/LEEF_2.archived.data/LEEF/3.archived.data/extracted/",
  output_dir = ".",
  format = "html"
)
}
\arguments{
\item{format}{the format of the report as a character vector of length 1.
Supported are at the moment: \code{html} (the default), \code{pdf} and \code{word}.}

\item{db}{fully qualified path to the sqlite database. The report will be saved to the same directory.}

\item{template}{Template to be used for report. At the moment only \code{"LEEF_1} and \code{"LEEF_2} supported.}

\item{suffix}{suffix for the file name}

\item{lastDays}{number of last days to be included in graph 2. Default: 7}
}
\value{
the fully qualified file name to the created report.
}
\description{
Report for checking pipeline and sampling
}
