% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_overlays.R
\name{update_overlays}
\alias{update_overlays}
\title{Just a function to update the overlays.}
\usage{
update_overlays(
  params = list(cores = 7, input_dir = "~/reclassification/", archive_dir =
    "/Volumes/LEEF/", extracted_dir =
    "/Volumes/LEEF/LEEF.archived.data/LEEF/3.archived.data/extracted/", avi_url =
    "https://cloud.s3it.uzh.ch:8080/v1/AUTH_0ac1f146d16c4aa8aee335872ef84aed/LEEF.archived.data/LEEF/3.archived.data/pre_processed/")
)
}
\arguments{
\item{extracted_dir}{directory with extracted data to be used}

\item{avi_url}{url in which all avis are in the subdirectories.}
}
\description{
Just a function to update the overlays.
}
