% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen_deposit.R
\name{zen_deposit}
\alias{zen_deposit}
\title{Deposit data on Zenodo}
\usage{
zen_deposit(
  token = NULL,
  sandbox = TRUE,
  archive_dir = "~/Duck/LEEFSwift3",
  timestamp,
  measuring_method,
  description = "Description of the data"
)
}
\arguments{
\item{token}{Zenodo token to upload the deposit}

\item{sandbox}{if \code{TRUE} (default) upload to the Zenodo sandbox for testing, if \code{FALSE} upload to the "real" Zenodo}

\item{archive_dir}{root directory of the archive directories}

\item{timestamp}{timestamp to be uploaded}

\item{measuring_method}{measuring method to be uploaded. Allowed values at the moment:
\itemize{
\item \code{"bemovi.mag.16"}
\item \code{"bemovi.mag.25"}
\item \code{"flowcam"}
\item \code{"flowcytometer"}
\item \code{"o2meter"}
\item \code{"manualcount"}
}}

\item{description}{Description of the two data deposits}
}
\value{
list with all info concerning the deposits
}
\description{
Deposit data on Zenodo
}
